Move partner referral links into a dedicated component.
